requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
deepchem==2.8.0
keras==3.3.3
tensorflow==2.16.1
joblib==1.4.2
rdkit-pypi==2024.3.5

